chore: register new global func

+++ packages/q3-api/lib/chores/onQueueCleanup@daily.js
@@ -1,14 +1,22 @@
 const { __$db: Model } = require('q3-core-scheduler');
 const moment = require('moment');
+const { lowerCase } = require('lodash');
 
 module.exports = async () => {
-  const [interval, unit] = String(
-    process.env.QUEUING_COLLECTION || '1_YEAR',
+  let [interval, unit] = String(
+    process.env.QUEUING_LIFETIME || '1_YEARS',
   ).split('_');
 
-  await Model.deleteMany({
-    due: {
-      $lt: moment().subtract(interval, unit).toDate(),
-    },
-  });
+  interval = Number(interval);
+  unit = lowerCase(unit);
+
+  if (
+    !Number.isNaN(interval) &&
+    ['years', 'months', 'weeks', 'days'].includes(unit)
+  )
+    await Model.deleteMany({
+      due: {
+        $lt: moment().subtract(interval, unit).toDate(),
+      },
+    });
 };
